requirements.txt
langchain 
langchain-openai 
openai 
chromadb 
gradio 
python-dotenv 
tiktoken 
langchain-community
-e .